requirements.txt
# Django Backend Requirements for Aqua-Racine
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
Pillow>=10.0
django-environ>=0.11
django-filter>=23.3
drf-spectacular>=0.26
django-jazzmin>=2.6
gunicorn>=21.2
whitenoise>=6.6
django-ckeditor>=6.7
dj-database-url>=2.1
psycopg2-binary>=2.9
